annotator/requirements.txt
absl-py==0.7.1
alabaster==0.7.8
apache-libcloud==2.4.0
applicationinsights==0.11.7
asn1crypto==0.24.0
astor==0.8.0
astroid==1.6.0
attrs==19.1.0
Babel==2.4.0
backcall==0.1.0
beautifulsoup4==4.8.0
bleach==3.1.0
boto==2.49.0
boto3==1.9.174
botocore==1.12.174
Brlapi==0.6.6
Brotli==1.0.7
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cloudpickle==0.5.2
command-not-found==0.3
conllu==1.3.2
coverage==4.5.3
cryptography==2.1.4
cupshelpers==1.0
decorator==4.4.0
defer==1.0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
dj-database-url==0.5.0
Django==2.1.7
django-cloud-browser==0.5.0
django-filter==2.0.0
django-heroku==0.3.1
django-polymorphic==2.0.3
django-pyodbc-azure==2.1.0.0
django-rest-polymorphic==0.1.8
django-webpack-loader==0.6.0
django-widget-tweaks==1.4.2
djangorestframework==3.8.2
djangorestframework-csv==2.1.0
djangorestframework-filters==0.10.2
djangorestframework-xml==1.4.0
docutils==0.14
entrypoints==0.3
environs==4.1.0
Faker==0.9.1
flake8==3.6.0
Flask==1.0.3
furl==2.0.0
gast==0.2.2
gensim==3.7.3
glove-python==0.1.0
google-pasta==0.1.7
grpcio==1.23.0
gunicorn==19.9.0
h5py==2.9.0
html5lib==0.999999999
httplib2==0.9.2
idna==2.8
imagesize==0.7.1
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itsdangerous==1.1.0
jedi==0.14.0
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jsonschema==3.0.1
jupyter-client==5.2.2
jupyter-core==4.4.0
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
lockfile==0.12.2
logilab-common==1.4.1
louis==3.5.0
lxml==4.3.4
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.1.1
MarkupSafe==1.1.1
marshmallow==3.0.0
mccabe==0.6.1
mistune==0.8.4
mixer==6.1.3
model-mommy==1.6.0
multidict==4.5.2
nbconvert==5.5.0
nbformat==4.4.0
netifaces==0.10.4
notebook==5.7.8
numpy==1.17.0
numpydoc==0.7.0
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
orderedmultidict==1.0.1
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.0
pbr==5.4.3
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
praatio==3.8.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.1
psutil==5.4.2
psycopg2==2.8.3
psycopg2-binary==2.7.7
ptyprocess==0.6.0
pycairo==1.16.2
pycodestyle==2.4.0
pycrypto==2.6.1
pycups==1.9.73
pyflakes==2.0.0
Pygments==2.4.2
pygobject==3.26.1
PyJWT==1.7.1
pylint==1.8.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyodbc==4.0.27
pyRFC3339==1.0
pyrsistent==0.15.2
python-apt==1.6.4
python-dateutil==2.7.3
python-debian==0.1.32
python-docx==0.8.10
python-dotenv==0.10.3
python3-openid==3.1.0
pytz==2018.4
pyxdg==0.25
PyYAML==5.1.2
pyzmq==18.0.1
QtAwesome==0.4.4
qtconsole==4.5.1
QtPy==1.3.1
reportlab==3.4.0
requests==2.21.0
requests-oauthlib==1.2.0
requests-unixsocket==0.1.5
roman==2.0.0
rope==0.10.5
s3transfer==0.2.1
scikit-learn==0.21.2
scipy==1.3.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.5.0
seqeval==0.0.6
simplegeneric==0.8.1
simplejson==3.13.2
six==1.11.0
sklearn==0.0
smart-open==1.8.4
social-auth-app-django==3.1.0
social-auth-core==3.0.0
soupsieve==1.9.3
Sphinx==1.6.7
spyder==3.2.6
ssh-import-id==5.7
style==1.1.0
system-service==0.3
systemd-python==234
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
testresources==2.0.1
text-unidecode==1.2
torch==1.1.0
torchvision==0.3.0
tornado==5.0.2
traitlets==4.3.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unicodecsv==0.14.1
unittest-xml-reporting==2.5.1
update==0.0.1
urllib3==1.24.3
usb-creator==0.3.3
vcrpy==2.0.1
vcrpy-unittest==0.1.7
virtualenv==16.6.1
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
whitenoise==4.1.2
widgetsnbextension==3.4.2
wrapt==1.11.2
xkit==0.0.0
yarl==1.3.0
zope.interface==4.3.2
